Feat: Implement multilingual support (english and spanish). Now the js files are separated
