+++ public/translations/es.json
@@ -1,0 +1,16 @@
+{
+    "Roulette": "Ruleta",
+    "Spin the wheel": "Gira la ruleta",
+    "Instructions": "Instrucciones",
+    "Add, remove, or modify the options you want in the white box.": "Agrega, elimina o modifica las opciones que desees en el recuadro blanco.",
+    "Click on the wheel to start it spinning.": "Haz un click encima de la ruleta para que esta empiece a girar.",
+    "An option will be randomly selected and displayed on the screen.": "De manera aleatoria se seleccionará una opción y se mostrará en pantalla.",
+    "View on GitHub": "Ver proyecto en GitHub",
+    "Do you want to contact me?": "¿Quieres contactarme?",
+    "Email": "Correo",
+    "Message": "Mensaje",
+    "Send": "Enviar",
+    "Created by:": "Creado por:",
+    "Selected option": "Opción seleccionada",
+    "Remove option": "Borrar opción"
+}
